Hack ExecuteProgrammData to work with latest EDU scheduler
The scheduler already expects a 4-byte timestamp instead of a 2-byte
queue ID.

+++ Sts1CobcSw/Periphery/EduStructs.hpp
@@ -39,7 +39,7 @@ struct ExecuteProgramData
 {
     static constexpr auto id = executeProgramId;
     ts::uint16_t programId;
-    ts::uint16_t queueId;
+    ts::uint32_t queueId;
     ts::int16_t timeout;
 };
